Push file data source behind an interface.
The interface allows to swap between different sources easily,
without having to know what kind of data source is in use, eg.
we are using a resampler to start with, but if we record on a
track it switches to a passthrough, and since the interface is
the same, no code changes are needed to switch.

This should make it easier/possible to have different sources on
different tracks, like track 1 is just a straight sample player,
track 2 resamples to loop properly, track 3 has pitch/time
adjustments on-load.